deployment/cdk/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Python utilities
boto3>=1.26.0

# Additional CDK modules for frontend deployment
aws-cdk.aws-s3-deployment>=2.100.0
aws-cdk.aws-cloudfront>=2.100.0
aws-cdk.aws-cloudfront-origins>=2.100.0
